+++ play/src/main/java/com/example/play/TrackInterceptor.kt
@@ -1,5 +1,5 @@
 package com.example.play
 
 interface TrackInterceptor {
-    fun onStart(duration: Double): DoubleArray
+    fun onStart(duration: Double): LongArray
 }
